fixed a few C warnings

+++ espresso-src/contain.c
@@ -148,7 +148,7 @@ pset_family dist_merge(pset_family A, pset mask)
 */
 pset_family d1merge(pset_family A, int var)
                                 /* disposes of A */
-           
+
 {
     return dist_merge(A, cube.var_mask[var]);
 }
@@ -157,23 +157,23 @@ pset_family d1merge(pset_family A, int var)
 
 /* d1_rm_equal -- distance-1 merge (merge cubes which are equal under a mask) */
 int d1_rm_equal(register pset *A1, qsort_compare_func compare)
-                  				/* array of set pointers */
-                 				/* comparison function */
+                          /* array of set pointers */
+                         /* comparison function */
 {
     register int i, j, dest;
 
     dest = 0;
     if (A1[0] != (pcube) NULL) {
-	for(i = 0, j = 1; A1[j] != (pcube) NULL; j++)
-	    if ( (*compare)(&A1[i], &A1[j]) == 0) {
-		/* if sets are equal (under the mask) merge them */
-		set_or(A1[i], A1[i], A1[j]);
-	    } else {
-		/* sets are unequal, so save the set i */
-		A1[dest++] = A1[i];
-		i = j;
-	    }
-	A1[dest++] = A1[i];
+  for(i = 0, j = 1; A1[j] != (pcube) NULL; j++)
+      if ( (*compare)(&A1[i], &A1[j]) == 0) {
+    /* if sets are equal (under the mask) merge them */
+    set_or(A1[i], A1[i], A1[j]);
+      } else {
+    /* sets are unequal, so save the set i */
+    A1[dest++] = A1[i];
+    i = j;
+      }
+  A1[dest++] = A1[i];
     }
     A1[dest] = (pcube) NULL;
     return dest;
@@ -183,16 +183,16 @@ int d1_rm_equal(register pset *A1, qsort_compare_func compare)
 /* rm_equal -- scan a sorted array of set pointers for duplicate sets */
 int rm_equal(pset *A1, qsort_compare_func compare)
                                 /* updated in place */
-                    
+
 {
     register pset *p, *pdest = A1;
 
     if (*A1 != NULL) {                  /* If more than one set */
-	for(p = A1+1; *p != NULL; p++)
-	    if ((*compare)(p, p-1) != 0)
-		*pdest++ = *(p-1);
-	*pdest++ = *(p-1);
-	*pdest = NULL;
+  for(p = A1+1; *p != NULL; p++)
+      if ((*compare)(p, p-1) != 0)
+    *pdest++ = *(p-1);
+  *pdest++ = *(p-1);
+  *pdest = NULL;
     }
     return pdest - A1;
 }
@@ -208,16 +208,16 @@ int rm_contain(pset *A1)
 
     /* Loop for all cubes of A1 */
     for(pa = A1; (a = *pa++) != NULL; ) {
-	/* Update the check pointer if the size has changed */
-	if (SIZE(a) != last_size)
-	    last_size = SIZE(a), pcheck = pdest;
-	for(pb = A1; pb != pcheck; ) {
-	    b = *pb++;
-	    INLINEsetp_implies(a, b, /* when_false => */ continue);
-	    goto lnext1;
-	}
-	/* set a was not contained by some larger set, so save it */
-	*pdest++ = a;
+  /* Update the check pointer if the size has changed */
+  if ((int)SIZE(a) != last_size)
+      last_size = SIZE(a), pcheck = pdest;
+  for(pb = A1; pb != pcheck; ) {
+      b = *pb++;
+      INLINEsetp_implies(a, b, /* when_false => */ continue);
+      goto lnext1;
+  }
+  /* set a was not contained by some larger set, so save it */
+  *pdest++ = a;
     lnext1: ;
     }
 
@@ -236,16 +236,16 @@ int rm_rev_contain(pset *A1)
 
     /* Loop for all cubes of A1 */
     for(pa = A1; (a = *pa++) != NULL; ) {
-	/* Update the check pointer if the size has changed */
-	if (SIZE(a) != last_size)
-	    last_size = SIZE(a), pcheck = pdest;
-	for(pb = A1; pb != pcheck; ) {
-	    b = *pb++;
-	    INLINEsetp_implies(b, a, /* when_false => */ continue);
-	    goto lnext1;
-	}
-	/* the set a did not contain some smaller set, so save it */
-	*pdest++ = a;
+  /* Update the check pointer if the size has changed */
+  if ((int)SIZE(a) != last_size)
+      last_size = SIZE(a), pcheck = pdest;
+  for(pb = A1; pb != pcheck; ) {
+      b = *pb++;
+      INLINEsetp_implies(b, a, /* when_false => */ continue);
+      goto lnext1;
+  }
+  /* the set a did not contain some smaller set, so save it */
+  *pdest++ = a;
     lnext1: ;
     }
 
@@ -257,27 +257,27 @@ int rm_rev_contain(pset *A1)
 /* rm2_equal -- check two sorted arrays of set pointers for equal cubes */
 int rm2_equal(register pset *A1, register pset *B1, pset *E1, qsort_compare_func compare)
                                         /* updated in place */
-             
-                    
+
+
 {
     register pset *pda = A1, *pdb = B1, *pde = E1;
 
     /* Walk through the arrays advancing pointer to larger cube */
     for(; *A1 != NULL && *B1 != NULL; )
-	switch((*compare)(A1, B1)) {
-	    case -1:    /* "a" comes before "b" */
-		*pda++ = *A1++; break;
-	    case 0:     /* equal cubes */
-		*pde++ = *A1++; B1++; break;
-	    case 1:     /* "a" is to follow "b" */
-		*pdb++ = *B1++; break;
-	}
+  switch((*compare)(A1, B1)) {
+      case -1:    /* "a" comes before "b" */
+    *pda++ = *A1++; break;
+      case 0:     /* equal cubes */
+    *pde++ = *A1++; B1++; break;
+      case 1:     /* "a" is to follow "b" */
+    *pdb++ = *B1++; break;
+  }
 
     /* Finish moving down the pointers of A and B */
     while (*A1 != NULL)
-	*pda++ = *A1++;
+  *pda++ = *A1++;
     while (*B1 != NULL)
-	*pdb++ = *B1++;
+  *pdb++ = *B1++;
     *pda = *pdb = *pde = NULL;
 
     return pde - E1;
@@ -293,14 +293,14 @@ int rm2_contain(pset *A1, pset *B1)
 
     /* for each set in the first array ... */
     for(pa = A1; (a = *pa++) != NULL; ) {
-	/* for each set in the second array which is larger ... */
-	for(pb = B1; (b = *pb++) != NULL && SIZE(b) > SIZE(a); ) {
-	    INLINEsetp_implies(a, b, /* when_false => */ continue);
-	    /* set was contained in some set of B, so don't save pointer */
-	    goto lnext1;
-	}
-	/* set wasn't contained in any set of B, so save the pointer */
-	*pdest++ = a;
+  /* for each set in the second array which is larger ... */
+  for(pb = B1; (b = *pb++) != NULL && SIZE(b) > SIZE(a); ) {
+      INLINEsetp_implies(a, b, /* when_false => */ continue);
+      /* set was contained in some set of B, so don't save pointer */
+      goto lnext1;
+  }
+  /* set wasn't contained in any set of B, so save the pointer */
+  *pdest++ = a;
     lnext1: ;
     }
 
@@ -318,8 +318,8 @@ pset *sf_sort(pset_family A, qsort_compare_func compare)
     /* Create a single array pointing to each cube of A */
     pdest = A1 = ALLOC(pset, A->count + 1);
     foreach_set(A, last, p) {
-	PUTSIZE(p, set_ord(p));         /* compute the set size */
-	*pdest++ = p;                   /* save the pointer */
+  PUTSIZE(p, set_ord(p));         /* compute the set size */
+  *pdest++ = p;                   /* save the pointer */
     }
     *pdest = NULL;                      /* Sentinel -- never seen by sort */
 
@@ -337,7 +337,7 @@ pset *sf_list(register pset_family A)
     /* Create a single array pointing to each cube of A */
     pdest = A1 = ALLOC(pset, A->count + 1);
     foreach_set(A, last, p)
-	*pdest++ = p;                   /* save the pointer */
+  *pdest++ = p;                   /* save the pointer */
     *pdest = NULL;                      /* Sentinel */
     return A1;
 }
@@ -351,7 +351,7 @@ pset_family sf_unlist(pset *A1, int totcnt, int size)
 
     R->count = totcnt;
     for(pr = R->data, pa = A1; (p = *pa++) != NULL; pr += R->wsize)
-	INLINEset_copy(pr, p);
+  INLINEset_copy(pr, p);
     FREE(A1);
     return R;
 }
@@ -367,11 +367,11 @@ pset_family sf_ind_unlist(pset *A1, int totcnt, int size, int *row_indices, regi
     R->count = totcnt;
     new_row_indices = ALLOC(int, totcnt);
     for(pr = R->data, pa = A1, i=0; (p = *pa++) != NULL; pr += R->wsize, i++) {
-	INLINEset_copy(pr, p);
-	new_row_indices[i] = row_indices[(p - pfirst)/R->wsize];
+  INLINEset_copy(pr, p);
+  new_row_indices[i] = row_indices[(p - pfirst)/R->wsize];
     }
     for(i = 0; i < totcnt; i++)
-	row_indices[i] = new_row_indices[i];
+  row_indices[i] = new_row_indices[i];
     FREE(new_row_indices);
     FREE(A1);
     return R;
@@ -381,7 +381,7 @@ pset_family sf_ind_unlist(pset *A1, int totcnt, int size, int *row_indices, regi
 /* sf_merge -- merge three sorted lists of set pointers */
 pset_family sf_merge(pset *A1, pset *B1, pset *E1, int totcnt, int size)
                                         /* will be disposed of */
-                    
+
 {
     register pset pr, ps, *pmin, *pmid, *pmax;
     pset_family R;
@@ -396,24 +396,24 @@ pset_family sf_merge(pset *A1, pset *B1, pset *E1, int totcnt, int size)
     /* Quick bubble sort to order the top member of the three arrays */
     n = 3;  temp[0] = A1;  temp[1] = B1;  temp[2] = E1;
     for(i = 0; i < n-1; i++)
-	for(j = i+1; j < n; j++)
-	    if (desc1(*temp[i], *temp[j]) > 0) {
-		swap = temp[j];
-		temp[j] = temp[i];
-		temp[i] = swap;
-	    }
+  for(j = i+1; j < n; j++)
+      if (desc1(*temp[i], *temp[j]) > 0) {
+    swap = temp[j];
+    temp[j] = temp[i];
+    temp[i] = swap;
+      }
     pmin = temp[0];  pmid = temp[1];  pmax = temp[2];
 
     /* Save the minimum element, then update pmin, pmid, pmax */
     while (*pmin != (pset) NULL) {
-	ps = *pmin++;
-	INLINEset_copy(pr, ps);
-	pr += R->wsize;
-	if (desc1(*pmin, *pmax) > 0) {
-	    swap = pmax; pmax = pmin; pmin = pmid; pmid = swap;
-	} else if (desc1(*pmin, *pmid) > 0) {
-	    swap = pmin; pmin = pmid; pmid = swap;
-	}
+  ps = *pmin++;
+  INLINEset_copy(pr, ps);
+  pr += R->wsize;
+  if (desc1(*pmin, *pmax) > 0) {
+      swap = pmax; pmax = pmin; pmin = pmid; pmid = swap;
+  } else if (desc1(*pmin, *pmid) > 0) {
+      swap = pmin; pmin = pmid; pmid = swap;
+  }
     }
 
     FREE(A1);
